requirements.txt
aiogram==2.11.2
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
certifi==2020.11.8
chardet==3.0.4
idna==2.10
multidict==5.0.2
psycopg2==2.8.6
python-dotenv==0.15.0
pytz==2020.4
SQLAlchemy==1.3.20
typing-extensions==3.7.4.3
yarl==1.6.3
gino~=1.0.1
envparse~=0.2.0
ConfigArgParse~=1.2.3
alembic~=1.4.3